+++ internal/otelcollector/config/builder/metric/gateway/config_builder.go
@@ -68,74 +68,6 @@ func makeReceiversConfig() config.ReceiversConfig {
 	}
 }
 
-func makeProcessorsConfig() config.ProcessorsConfig {
-	k8sAttributes := []string{
-		"k8s.pod.name",
-		"k8s.node.name",
-		"k8s.namespace.name",
-		"k8s.deployment.name",
-		"k8s.statefulset.name",
-		"k8s.daemonset.name",
-		"k8s.cronjob.name",
-		"k8s.job.name",
-	}
-
-	podAssociations := []config.PodAssociations{
-		{
-			Sources: []config.PodAssociation{
-				{
-					From: "resource_attribute",
-					Name: "k8s.pod.ip",
-				},
-			},
-		},
-		{
-			Sources: []config.PodAssociation{
-				{
-					From: "resource_attribute",
-					Name: "k8s.pod.uid",
-				},
-			},
-		},
-		{
-			Sources: []config.PodAssociation{
-				{
-					From: "connection",
-				},
-			},
-		},
-	}
-	return config.ProcessorsConfig{
-		Batch: &config.BatchProcessorConfig{
-			SendBatchSize:    1024,
-			Timeout:          "10s",
-			SendBatchMaxSize: 1024,
-		},
-		MemoryLimiter: &config.MemoryLimiterConfig{
-			CheckInterval:        "1s",
-			LimitPercentage:      75,
-			SpikeLimitPercentage: 10,
-		},
-		K8sAttributes: &config.K8sAttributesProcessorConfig{
-			AuthType:    "serviceAccount",
-			Passthrough: false,
-			Extract: config.ExtractK8sMetadataConfig{
-				Metadata: k8sAttributes,
-			},
-			PodAssociation: podAssociations,
-		},
-		Resource: &config.ResourceProcessorConfig{
-			Attributes: []config.AttributeAction{
-				{
-					Action: "insert",
-					Key:    "k8s.cluster.name",
-					Value:  "${KUBERNETES_SERVICE_HOST}",
-				},
-			},
-		},
-	}
-}
-
 func makePipelineConfig(outputAliases []string) config.PipelineConfig {
 	return config.PipelineConfig{
 		Receivers:  []string{"otlp"},
